requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
Django==3.2.7
django-heroku
djangorestframework
Pillow==8.3.2
pytz==2021.1
sqlparse==0.4.2
whitenoise==5.3.0
